Implement routing and refactor components in seperate views
