+++ client/index.html
@@ -1,10 +1,7 @@
 <!DOCTYPE html>
 <html>
   <head>
-
     <!-- React -->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/react/0.13.3/react.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/react/0.13.3/JSXTransformer.js"></script>
     <script src="https://code.jquery.com/jquery-2.1.4.js"></script>
 
     <!-- Bootstrap -->
