+++ client/src/app.jsx
@@ -1,13 +1,61 @@
-var GameBox = require('./views/gameBox.jsx');
+var React = require('react');
+
+var OverView = require('./views/OverView.jsx');
+var DetailView = require('./views/DetailView.jsx');
+
+var Router = require('react-router');
+var RouteHandler = Router.RouteHandler;
+var DefaultRoute = Router.DefaultRoute;
+var Route = Router.Route;
+
 
 var App = React.createClass({
+  getInitialState: function(){
+    return {
+      questions: []
+    };
+  },
+
+  loadAllQuestions: function(){
+    $.ajax({
+      url: 'http://localhost:3000/questions',
+      method: 'GET',
+      dataType: 'json',
+      success: function(data){
+        data.sort(function(a, b){
+          return a.qNumber - b.qNumber;
+        });
+        this.setState({questions: data});
+      }.bind(this),
+      error: function(xhr, status, err){
+        console.error(xhr, status, err.message);
+      }
+    });
+  },
+
+  componentDidMount: function(){
+    this.loadAllQuestions();
+  },
+
   render: function() {
     return (
       <div className="container">
-        <GameBox/>
+        <h2 className="title">Regex Game</h2>
+        <RouteHandler questions={this.state.questions}/>
       </div>
     )
   }
+
 });
 
-React.render(<App/>, document.body);
+var routes = (
+  <Route name="app" path="/" handler={App}>
+    <Route name="question" path="/:qNumber" handler={DetailView}/>
+    <DefaultRoute name="default" handler={OverView} />
+  </Route>
+);
+
+Router.run(routes, function(Root){
+  React.render(<Root />, document.body);
+});
+
