+++ package.json
@@ -20,7 +20,9 @@
   "dependencies": {
     "body-parser": "^1.13.3",
     "express": "^4.13.3",
-    "mongoose": "^4.1.1"
+    "mongoose": "^4.1.1",
+    "react": "^0.13.3",
+    "react-router": "^0.13.3"
   },
   "devDependencies": {
     "chai": "^3.2.0",
